Rename gitlab-cli.ts to gitlab.ts
- the file will be split into multiple files in the future
